+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Upcoming
 
 * Add GvG tab (showing sector army details)
+* Fix GB calculation when self (still unsafe) is followed by owner
 
 # 0.3.2
 * Fix battle info for newest version of FoE